Refactor Skeleton to use BaseSkeleton component
Extracted the core skeleton logic into a new BaseSkeleton component for both web and native platforms. Updated the Skeleton index to use BaseSkeleton, simplifying the code and improving maintainability. Introduced SkeletonProvider and related context utilities for group loading state.
