+++ packages/components/src/primitives/Skeleton/const.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_SKELETON_SIZE = 32;
+export const DEFAULT_RADIUS = 8;
