+++ packages/components/src/primitives/Skeleton/context.ts
@@ -1,0 +1,12 @@
+import { createContext, useContext } from 'react';
+
+export const SkeletonProvider = createContext<{
+  isLoading: boolean | undefined;
+}>({
+  isLoading: undefined,
+});
+
+export const useIsGroupLoading = () => {
+  const { isLoading } = useContext(SkeletonProvider);
+  return isLoading;
+};
