+++ packages/components/src/primitives/Skeleton/index.tsx
@@ -1,96 +1,21 @@
-import { createContext, forwardRef, useContext, useMemo } from 'react';
+import { forwardRef, useMemo } from 'react';
 
-import { SkeletonView } from '@onekeyfe/react-native-skeleton';
-
-import type { StackStyle } from '@onekeyhq/components/src/shared/tamagui';
 import {
   styled,
-  usePropsAndStyle,
-  useThemeName,
   withStaticProperties,
 } from '@onekeyhq/components/src/shared/tamagui';
 
-import { Stack, YStack } from '../Stack';
+import { YStack } from '../Stack';
 
-const SkeletonProvider = createContext<{
-  isLoading: boolean | undefined;
-}>({
-  isLoading: undefined,
-});
+import { BaseSkeleton } from './BaseSkeleton';
+import { SkeletonProvider } from './context';
 
-const useIsGroupLoading = () => {
-  const { isLoading } = useContext(SkeletonProvider);
-  return isLoading;
-};
-
-export type ISkeletonProps = StackStyle & {
-  isLoading?: boolean;
-  radius?: 'round' | 'square' | number;
-  colorMode?: 'dark' | 'light';
-  children?: React.ReactNode;
-};
-
-const DEFAULT_SKELETON_SIZE = 32;
-const DEFAULT_RADIUS = 8;
-const baseColors = {
-  dark: ['#111111', '#333333'],
-  light: ['#fafafa', '#cdcdcd'],
-};
-function BasicSkeleton({
-  isLoading = false,
-  colorMode,
-  children,
-  ...props
-}: ISkeletonProps) {
-  const [restProps, style] = usePropsAndStyle(props, {
-    resolveValues: 'auto',
-  });
-  const themeName = useThemeName();
-  const colors =
-    (colorMode ?? themeName) === 'dark' ? baseColors.dark : baseColors.light;
-
-  const borderRadius = useMemo(() => {
-    if (restProps.radius === 'round') {
-      return 9999;
-    }
-    if (restProps.radius === 'square') {
-      return 0;
-    }
-    return restProps.radius || DEFAULT_RADIUS;
-  }, [restProps.radius]);
-
-  const isGroupLoading = useIsGroupLoading();
-  return isGroupLoading === undefined || isGroupLoading ? (
-    <Stack
-      bg="$bg"
-      style={style as any}
-      height={style.height || DEFAULT_SKELETON_SIZE}
-      width={style.width || '100%'}
-      borderRadius={borderRadius}
-      overflow="hidden"
-      {...restProps}
-    >
-      <SkeletonView
-        style={[
-          style as any,
-          {
-            height: style.height || DEFAULT_SKELETON_SIZE,
-            width: style.width || '100%',
-          },
-        ]}
-        shimmerSpeed={3}
-        shimmerGradientColors={colors}
-      />
-    </Stack>
-  ) : (
-    children || null
-  );
-}
+import type { ISkeletonProps } from './type';
 
 function BodySmSkeleton({ ...props }: ISkeletonProps) {
   return (
     <YStack py="$1">
-      <BasicSkeleton w={64} h="$2" {...props} />
+      <BaseSkeleton w={64} h="$2" {...props} />
     </YStack>
   );
 }
@@ -98,7 +23,7 @@ function BodySmSkeleton({ ...props }: ISkeletonProps) {
 function BodyMdSkeleton({ ...props }: ISkeletonProps) {
   return (
     <YStack py="$1">
-      <BasicSkeleton w={72} h="$3" {...props} />
+      <BaseSkeleton w={72} h="$3" {...props} />
     </YStack>
   );
 }
@@ -106,7 +31,7 @@ function BodyMdSkeleton({ ...props }: ISkeletonProps) {
 function BodyLgSkeleton({ ...props }: ISkeletonProps) {
   return (
     <YStack py="$1">
-      <BasicSkeleton w={82} h="$4" {...props} />
+      <BaseSkeleton w={82} h="$4" {...props} />
     </YStack>
   );
 }
@@ -114,7 +39,7 @@ function BodyLgSkeleton({ ...props }: ISkeletonProps) {
 function HeadingXsSkeleton({ ...props }: ISkeletonProps) {
   return (
     <YStack py="$1">
-      <BasicSkeleton w={91} h="$2" {...props} />
+      <BaseSkeleton w={91} h="$2" {...props} />
     </YStack>
   );
 }
@@ -122,7 +47,7 @@ function HeadingXsSkeleton({ ...props }: ISkeletonProps) {
 function HeadingSmSkeleton({ ...props }: ISkeletonProps) {
   return (
     <YStack py="$1">
-      <BasicSkeleton w={77} h="$3" {...props} />
+      <BaseSkeleton w={77} h="$3" {...props} />
     </YStack>
   );
 }
@@ -130,7 +55,7 @@ function HeadingSmSkeleton({ ...props }: ISkeletonProps) {
 function HeadingMdSkeleton({ ...props }: ISkeletonProps) {
   return (
     <YStack py="$1">
-      <BasicSkeleton w={86} h="$4" {...props} />
+      <BaseSkeleton w={86} h="$4" {...props} />
     </YStack>
   );
 }
@@ -138,7 +63,7 @@ function HeadingMdSkeleton({ ...props }: ISkeletonProps) {
 function HeadingLgSkeleton({ ...props }: ISkeletonProps) {
   return (
     <YStack py="$1">
-      <BasicSkeleton w={96} h="$4" {...props} />
+      <BaseSkeleton w={96} h="$4" {...props} />
     </YStack>
   );
 }
@@ -146,7 +71,7 @@ function HeadingLgSkeleton({ ...props }: ISkeletonProps) {
 function HeadingXlSkeleton({ ...props }: ISkeletonProps) {
   return (
     <YStack py="$1">
-      <BasicSkeleton w={103} h="$5" {...props} />
+      <BaseSkeleton w={103} h="$5" {...props} />
     </YStack>
   );
 }
@@ -154,7 +79,7 @@ function HeadingXlSkeleton({ ...props }: ISkeletonProps) {
 function Heading2XlSkeleton({ ...props }: ISkeletonProps) {
   return (
     <YStack py="$1">
-      <BasicSkeleton w={123} h="$6" {...props} />
+      <BaseSkeleton w={123} h="$6" {...props} />
     </YStack>
   );
 }
@@ -162,7 +87,7 @@ function Heading2XlSkeleton({ ...props }: ISkeletonProps) {
 function Heading3XlSkeleton({ ...props }: ISkeletonProps) {
   return (
     <YStack py="$1">
-      <BasicSkeleton w={143} h="$7" {...props} />
+      <BaseSkeleton w={143} h="$7" {...props} />
     </YStack>
   );
 }
@@ -170,7 +95,7 @@ function Heading3XlSkeleton({ ...props }: ISkeletonProps) {
 function Heading4XlSkeleton({ ...props }: ISkeletonProps) {
   return (
     <YStack py="$1">
-      <BasicSkeleton w={163} h="$8" {...props} />
+      <BaseSkeleton w={163} h="$8" {...props} />
     </YStack>
   );
 }
@@ -178,7 +103,7 @@ function Heading4XlSkeleton({ ...props }: ISkeletonProps) {
 function Heading5XlSkeleton({ ...props }: ISkeletonProps) {
   return (
     <YStack py="$1">
-      <BasicSkeleton w={209} h="$10" {...props} />
+      <BaseSkeleton w={209} h="$10" {...props} />
     </YStack>
   );
 }
@@ -199,7 +124,7 @@ function SkeletonGroup({
 }
 
 export const Skeleton = withStaticProperties(
-  styled(forwardRef(BasicSkeleton), {
+  styled(forwardRef(BaseSkeleton), {
     name: 'Skeleton',
   } as const),
   {
@@ -227,5 +152,7 @@ export function SkeletonContainer({
   isLoading: boolean;
   children: React.ReactNode;
 }) {
-  return isLoading ? <Skeleton {...props} isLoading /> : children;
+  return isLoading ? <Skeleton {...props} /> : children;
 }
+
+export type { ISkeletonProps } from './type';
