+++ packages/components/src/primitives/Skeleton/type.ts
@@ -1,0 +1,7 @@
+import type { StackStyle } from '@onekeyhq/components/src/shared/tamagui';
+
+export type ISkeletonProps = StackStyle & {
+  radius?: 'round' | 'square' | number;
+  colorMode?: 'dark' | 'light';
+  children?: React.ReactNode;
+};
